% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a table from a data_set}
\usage{
normalize(filename, nid = 4)
}
\arguments{
\item{nid}{The number of factors above which to normalize.}

\item{Filename}{The name of the file to be normalized (must be in current directory).}
}
\value{
The normalized data frame.
}
\description{
Normalize a table from a data_set
}
\examples{
dataScrapingProject_normalize ("testing_data.csv", 5)
dataScrapingProject_normalize ("testing_data.csv")
}

